move compose validation out of quota package
